pyproject.toml
[project]
name = "projetagilite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.2",
    "django-ninja>=1.5.3",
]
